prjunnamed_netlist
Struct MemoryReadFlipFlop

pub struct MemoryReadFlipFlop {
    pub clock: ControlNet,
    pub clear: ControlNet,
    pub reset: ControlNet,
    pub enable: ControlNet,
    pub reset_over_enable: bool,
    pub clear_value: Const,
    pub reset_value: Const,
    pub init_value: Const,
    pub relations: Vec<MemoryPortRelation>,
}

A structure describing a synchronous read portโs control signals and behavior.

The behavior of a synchronous read port is mostly the same as an asynchronous read port feeding a flip-flop with the controls described in this structure. However, synchronous read ports have special behavior when a read coincides with a write to the same memory row. This behavior is selected by the relations field.

The fields other than relations have the same meaning as the corresponding fields in the FlipFlop structure. The width of the reset, clear, and init values must be equal to the data_len of the port.

Fields
clock: ControlNetclear: ControlNetreset: ControlNetenable: ControlNetreset_over_enable: boolclear_value: Constreset_value: Constinit_value: Constrelations: Vec<MemoryPortRelation>

The behavior of this read port during a simultaneous write by a given write port.

Each entry in this vector describes the behavior of this read port in relation to a write port with the same index in the write_ports vector of the containing memory.

Only ports sharing the same clock net can have defined relations โ this vector must contain MemoryPortRelation::Undefined for every write port where write_port.clock != read_port.clock.

If a given memory bit is simultanously written by more than one write port while being read, the read data is undefined (as is the value written to memory), regardless of the relations between the ports.

Implementations

impl MemoryReadFlipFlop

pub fn with_clock(self, clock: impl Into<ControlNet>) -> Self

pub fn with_clear(self, clear: impl Into<ControlNet>) -> Self

pub fn with_clear_value( self, clear: impl Into<ControlNet>, clear_value: impl Into<Const>, ) -> Self

pub fn with_reset(self, reset: impl Into<ControlNet>) -> Self

pub fn with_reset_value( self, reset: impl Into<ControlNet>, reset_value: impl Into<Const>, ) -> Self

pub fn with_enable(self, enable: impl Into<ControlNet>) -> Self

pub fn with_init(self, value: impl Into<Const>) -> Self

pub fn has_clock(&self) -> bool

pub fn has_enable(&self) -> bool

pub fn has_reset(&self) -> bool

pub fn has_reset_value(&self) -> bool

pub fn has_clear(&self) -> bool

pub fn has_clear_value(&self) -> bool

pub fn has_init_value(&self) -> bool

pub fn remap_reset_over_enable(&mut self, design: &Design)

pub fn remap_enable_over_reset(&mut self, design: &Design)